split into sep mod files
