+++ backend/sql/add_user.sql
@@ -1,3 +1,3 @@
-INSERT INTO osqpseed.users(email, fullname)
+INSERT INTO osqspeed.users(email, fullname)
 VALUES ($1, $2)
 RETURNING $table_fields;
