+++ backend/src/db.rs
@@ -1,0 +1,19 @@
+use deadpool_postgres::Client;
+use tokio_pg_mapper::FromTokioPostgresRow;
+
+use crate::{errors::MyError, models::User};
+
+pub async fn add_user(client: &Client, user_info: User) -> Result<User, MyError> {
+    let _stmt = include_str!("../sql/add_user.sql");
+    let _stmt = _stmt.replace("$table_fields", &User::sql_table_fields());
+    let stmt = client.prepare(&_stmt).await.unwrap();
+
+    client
+        .query(&stmt, &[&user_info.email, &user_info.fullname])
+        .await?
+        .iter()
+        .map(|row| User::from_row_ref(row).unwrap())
+        .collect::<Vec<User>>()
+        .pop()
+        .ok_or(MyError::NotFound) // more applicable for SELECTs
+}
