+++ backend/src/main.rs
@@ -1,97 +1,11 @@
-mod config {
-    use serde::Deserialize;
-    #[derive(Debug, Default, Deserialize)]
-    pub struct ExampleConfig {
-        pub server_addr: String,
-        pub pg: deadpool_postgres::Config,
-    }
-}
-
-mod models {
-    use serde::{Deserialize, Serialize};
-    use tokio_pg_mapper_derive::PostgresMapper;
-
-    #[derive(Deserialize, PostgresMapper, Serialize)]
-    #[pg_mapper(table = "users")] // singular 'user' is a keyword..
-    pub struct User {
-        pub email: String,
-        pub fullname: String,
-    }
-}
-
-mod errors {
-    use actix_web::{HttpResponse, ResponseError};
-    use deadpool_postgres::PoolError;
-    use derive_more::{Display, From};
-    use tokio_pg_mapper::Error as PGMError;
-    use tokio_postgres::error::Error as PGError;
-
-    #[derive(Display, From, Debug)]
-    pub enum MyError {
-        NotFound,
-        PGError(PGError),
-        PGMError(PGMError),
-        PoolError(PoolError),
-    }
-    impl std::error::Error for MyError {}
-
-    impl ResponseError for MyError {
-        fn error_response(&self) -> HttpResponse {
-            match *self {
-                MyError::NotFound => HttpResponse::NotFound().finish(),
-                MyError::PoolError(ref err) => {
-                    HttpResponse::InternalServerError().body(err.to_string())
-                }
-                _ => HttpResponse::InternalServerError().finish(),
-            }
-        }
-    }
-}
-
-mod db {
-    use deadpool_postgres::Client;
-    use tokio_pg_mapper::FromTokioPostgresRow;
-
-    use crate::{errors::MyError, models::User};
-
-    pub async fn add_user(client: &Client, user_info: User) -> Result<User, MyError> {
-        let _stmt = include_str!("../sql/add_user.sql");
-        let _stmt = _stmt.replace("$table_fields", &User::sql_table_fields());
-        let stmt = client.prepare(&_stmt).await.unwrap();
-
-        client
-            .query(&stmt, &[&user_info.email, &user_info.fullname])
-            .await?
-            .iter()
-            .map(|row| User::from_row_ref(row).unwrap())
-            .collect::<Vec<User>>()
-            .pop()
-            .ok_or(MyError::NotFound) // more applicable for SELECTs
-    }
-}
-
-mod handlers {
-    use actix_web::{web, Error, HttpResponse};
-    use deadpool_postgres::{Client, Pool};
-
-    use crate::{db, errors::MyError, models::User};
-
-    pub async fn add_user(
-        user: web::Json<User>,
-        db_pool: web::Data<Pool>,
-    ) -> Result<HttpResponse, Error> {
-        let user_info: User = user.into_inner();
-
-        let client: Client = db_pool.get().await.map_err(MyError::PoolError)?;
-
-        let new_user = db::add_user(&client, user_info).await?;
-
-        Ok(HttpResponse::Ok().json(new_user))
-    }
-}
+mod config;
+mod db;
+mod errors;
+mod handlers;
+mod models;
 
 use ::config::Config;
-use actix_web::{web, App, HttpServer};
+use actix_web::{middleware, web, App, HttpServer};
 use dotenv::dotenv;
 use handlers::add_user;
 use tokio_postgres::NoTls;
@@ -113,6 +27,7 @@ async fn main() -> std::io::Result<()> {
 
     let server = HttpServer::new(move || {
         App::new()
+            .wrap(middleware::DefaultHeaders::new().add(("Server", "osqspeed-api powered by Rust")))
             .app_data(web::Data::new(pool.clone()))
             .service(web::resource("/users").route(web::post().to(add_user)))
     })
