+++ backend/src/models.rs
@@ -1,0 +1,9 @@
+use serde::{Deserialize, Serialize};
+use tokio_pg_mapper_derive::PostgresMapper;
+
+#[derive(Deserialize, PostgresMapper, Serialize)]
+#[pg_mapper(table = "users")] // singular 'user' is a keyword..
+pub struct User {
+    pub email: String,
+    pub fullname: String,
+}
